CCBC-1672: handle case when packet gets discarded before IO calls completion handler
When the operation is getting retried, its buffers are getting relocated
to new pipeline, and removed from the old one. But if the write
operation still in the kernel and completion-based IO (e.g. windows
IOCP) is being used, the state of the buffer manager of the old pipeline
have to be adjusted.

Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/233237
Tested-by: Build Bot <build@couchbase.com>

+++ plugins/io/iocp/iocp_iops.c
@@ -370,7 +370,8 @@ static void iops_dtor(lcb_io_opt_t iobase)
     {
         iocp_sockdata_t *sd = LCB_LIST_ITEM(cur, iocp_sockdata_t, list);
 
-        IOCP_LOG(IOCP_WARN, "Leak detected in socket %p (%lu). Refcount=%d", sd, sd->sSocket, sd->refcount);
+        IOCP_LOG(IOCP_WARN, "Leak detected in socket %p (%lu). Refcount=%d", sd, (unsigned long)sd->sSocket,
+                 sd->refcount);
         if (sd->sSocket != INVALID_SOCKET) {
             closesocket(sd->sSocket);
             sd->sSocket = INVALID_SOCKET;
